The UI now starts the show and Ticker on a separate Thread (runAsync)
The show will be the the central controlling unit in the future. The Akai
and the keyboard will just be like TV remotes which change variables while
the process is running. The process itself can run on its own.
The communication from the UI to the show is done by queuing up runnables
which will be executed on the show's Thread.